bumped CC module to 0.1.15
0.1.15
encoded the ttml URL before adding to the YQL query.

bumped CC module to 0.1.14

0.1.14
changed the interface construction for KONtx.cc.engineInterface and KONtx.cc.useHardware. now we perform system level checks for hardware support each time the api is called instead of only once during construction of the api. this was causing failures on some systems that reported hardware support late. also this is required to support the new design of tvAPI in our v7 products which always report captioning support late.